Refactor SD card blocks into separate module.
