+++ src/modules/editor.js
@@ -37,6 +37,7 @@ import './blockly/generators/propc/oled';
 import './blockly/generators/propc/heb';
 import './blockly/generators/propc/procedures';
 import './blockly/generators/propc/s3';
+import './blockly/generators/propc/sd_card';
 import './blockly/generators/propc/sensors';
 import './blockly/generators/propc/variables';
